Use apply_action when apply=true
This fixes an issue where `apply = true` causes an invocation to do
nothing. In particular, `action.action.command` is not set, so nothing
is done. I assume this path was from an old version of NeoVIM?

This PR instead uses the `apply_action` function on the picker, instead.
It seems a bit odd to find a picker just to invoke this function, but my
Lua skills are insufficient to refactor further.